+++ src/redturtle/prenotazioni/restapi/serializers/adapters/prenotazioni_folder.py
@@ -1,12 +1,14 @@
 # -*- coding: utf-8 -*-
-from plone import api
 from plone.restapi.interfaces import ISerializeToJson
 from plone.restapi.serializer.dxcontent import SerializeFolderToJson
-from zope.component import adapter
+from zope.component import adapter, getMultiAdapter
 from zope.interface import implementer
 
 from redturtle.prenotazioni.content.prenotazioni_folder import IPrenotazioniFolder
-from redturtle.prenotazioni.interfaces import IRedturtlePrenotazioniLayer
+from redturtle.prenotazioni.interfaces import (
+    IRedturtlePrenotazioniLayer,
+    ISerializeToRetroCompatibleJson,
+)
 
 
 @implementer(ISerializeToJson)
@@ -15,12 +17,9 @@ class PrenotazioniFolderSerializer(SerializeFolderToJson):
     def __call__(self, *args, **kwargs):
         res = super().__call__()
 
-        if res.get("booking_types"):
-            if api.user.has_permission("redturtle.prenotazioni.ViewHiddenTypes"):
-                return res
-
-            res["booking_types"] = [
-                t for t in res["booking_types"] if not t.get("hidden")
-            ]
+        res["booking_types"] = [
+            getMultiAdapter((i, self.request), ISerializeToRetroCompatibleJson)()
+            for i in self.context.get_booking_types()
+        ]
 
         return res