+++ src/redturtle/prenotazioni/tests/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+# TODO: Move all the prenotazioni folder structure setups to one generic method