+++ src/redturtle/prenotazioni/vocabularies/voc_durata_incontro.py
@@ -20,7 +20,7 @@ class VocDurataIncontro(object):
     """ """
 
     def __call__(self, context):
-        items = [VocabItem(str(x), str(x)) for x in range(10, 95, 5)]
+        items = [VocabItem(x, x) for x in range(10, 185, 5)]
 
         if not IDexterityContent.providedBy(context):
             req = getRequest()
@@ -29,7 +29,11 @@ class VocDurataIncontro(object):
         terms = []
         for item in items:
             terms.append(
-                SimpleTerm(value=item.token, token=str(item.token), title=item.value)
+                SimpleTerm(
+                    value=str(item.token),
+                    token=str(item.token),
+                    title=item.value,
+                )
             )
         return SimpleVocabulary(terms)